% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{createGSEObj}
\alias{createGSEObj}
\title{Take a GEOQuery object and convert it to a MetaIntegrator gse.obj}
\usage{
createGSEObj(geoQueryObj)
}
\arguments{
\item{geoQueryObj}{that is the output of}
}
\value{
gse.obj with expression data and keys extracted
}
\description{
Take a GEOQuery object and convert it to a MetaIntegrator gse.obj
}
