% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectGeneID.R
\name{.check_entrez_overlap}
\alias{.check_entrez_overlap}
\title{Count the length of overlap with entrez ids for a sequence of cols.}
\usage{
.check_entrez_overlap(gpl.df, cols, organism, ref_dir = NULL)
}
\arguments{
\item{gpl.df}{the feature data from the GPL data frame}

\item{cols}{the columns you think contain entrez ids to check for overlap with}

\item{organism}{the organism: rat, mouse, or human}

\item{ref_dir}{option to set reference directory  (defaults to tempdir)}
}
\value{
the length of overlap for all the columns, where names are the column indices
}
\description{
Count the length of overlap with entrez ids for a sequence of cols.
}
