% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_lab_code.R
\name{.checkExprMatFormat}
\alias{.checkExprMatFormat}
\title{Checks that an expression matrix has the correct format}
\usage{
.checkExprMatFormat(expr_mat)
}
\arguments{
\item{expr_mat}{}
}
\description{
Checks that an expression matrix has the correct format
}
