% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkFitFormat}
\alias{.checkFitFormat}
\title{Checks that a fit object has the correct structure}
\usage{
.checkFitFormat(fit)
}
\arguments{
\item{fit}{a sex labeling fit with attributes "f" (list of f genes), "m" (list of m genes), and "threshold" (cutoff score)}
}
\description{
Checks that a fit object has the correct structure
}
