% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkGeneFormat}
\alias{.checkGeneFormat}
\title{Checks that the expression matrix rownames contain the genes of interest.}
\usage{
.checkGeneFormat(mat, genes)
}
\arguments{
\item{mat}{the expression matrix, rows should be gene names}

\item{genes}{list of genes}
}
\description{
Checks that the expression matrix rownames contain the genes of interest.
}
