% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkMatching}
\alias{.checkMatching}
\title{Checks that sex labels and expression matrix match}
\usage{
.checkMatching(expr_mat, sex_lab)
}
\arguments{
\item{sex_lab}{list of sex labels formatted as 0s/1s}
}
\description{
Checks that sex labels and expression matrix match
}
