% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkProbeMapping}
\alias{.checkProbeMapping}
\title{Checks that a probe mapping has the correct format.}
\usage{
.checkProbeMapping(probe_mat, probe_map, gene_list)
}
\arguments{
\item{probe_mat}{an expression matrix with probes as rows and columns as samples}

\item{probe_map}{list mapping from probes to genes, names are probes, values are genes}

\item{gene_list}{list of all genes to extract}
}
\description{
Briefly, the probe mapping should be formatted as a list with names as probes and values as genes.
The names of the probe map should match the rownames of the matrix and the values should match
the gene list.
This function looks at the intersections, and throws an error if there is no intersection
and prints a warning if the intersection is small (10% of either).
}
