% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkSexLabFormat}
\alias{.checkSexLabFormat}
\title{Checks that input sex labels have the correct format}
\usage{
.checkSexLabFormat(sex_lab)
}
\arguments{
\item{sex_lab}{list of sex labels formatted as 0s/1s}
}
\description{
Make sure the sex labels are all 0/1 (NAs ok, but not all)
}
