% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkTrainInputFormat}
\alias{.checkTrainInputFormat}
\title{Checks that expression matrix and sex labels have the correct structure for training}
\usage{
.checkTrainInputFormat(expr_mat, sex_lab)
}
\arguments{
\item{expr_mat}{the expression matrix, rows should be gene names, columns are samples}

\item{sex_lab}{list of sex labels, names of list should match column names}
}
\description{
Checks that expression matrix and sex labels have the correct structure for training
}
