% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGPL.R
\name{.clean_mapping}
\alias{.clean_mapping}
\title{Clean up the mapping output}
\usage{
.clean_mapping(mapped)
}
\arguments{
\item{mapped}{a data frame of mapped data}
}
\value{
cleaned mapping with NAs/empty strings removed
}
\description{
Clean up the mapping output
}
