% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{.convertGenes}
\alias{.convertGenes}
\title{Convert the expression matrix to genes.
This takes the average of probes to map them to a gene.}
\usage{
.convertGenes(expData, probe_gene)
}
\arguments{
\item{expData}{the expression matrix with rows as probes and columns as conditions}

\item{probe_gene}{a mapping frame with a probe and gene column}
}
\value{
an expression matrix with genes as rows
}
\description{
Convert the expression matrix to genes.
This takes the average of probes to map them to a gene.
}
