% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapDetectGeneID.R
\name{.detect_ensembl_cols}
\alias{.detect_ensembl_cols}
\title{Identify the ensembl columns and return the one with the max overlap.}
\usage{
.detect_ensembl_cols(gpl.df, organism, MIN.OVERLAP = 8000)
}
\arguments{
\item{gpl.df}{the gpl data frame}

\item{organism}{the organism we are looking at}

\item{MIN.OVERLAP}{the minimum number of mapped genes allowed, default=8000}
}
\value{
the index of the column containing these ids
}
\description{
Identify the ensembl columns and return the one with the max overlap.
}
