% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapDetectGeneID.R
\name{.detect_entrez_ids_name}
\alias{.detect_entrez_ids_name}
\title{Looks for entrez ids by trying to find a column name named "Entrez"}
\usage{
.detect_entrez_ids_name(gpl_cols)
}
\arguments{
\item{gpl_cols}{the column data from a GPL}
}
\value{
a list of columns containing these locations, an empty list if no columns
}
\description{
Looks for entrez ids by trying to find a column name named "Entrez"
}
