% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapDetectGeneID.R
\name{.detect_genbank_cols}
\alias{.detect_genbank_cols}
\title{Identify the genbank columns and return the first
(Note - this is by looking for a handful of ids...)}
\usage{
.detect_genbank_cols(gpl.df, organism)
}
\arguments{
\item{gpl.df}{the gpl data frame}

\item{organism}{the organism we are looking at}
}
\value{
the index of the column containing these ids
}
\description{
Identify the genbank columns and return the first
(Note - this is by looking for a handful of ids...)
}
