% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectGeneID.R
\name{.detect_genbank_cols_name}
\alias{.detect_genbank_cols_name}
\title{Looks for genbank cols by trying to find a column name named "GB_ACC"
or genbank}
\usage{
.detect_genbank_cols_name(gpl_cols)
}
\arguments{
\item{gpl_cols}{the column data from a GPL}
}
\value{
a list of columns containing these locations, an empty list if no columns
}
\description{
Looks for genbank cols by trying to find a column name named "GB_ACC"
or genbank
}
