% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectGeneID.R
\name{.detect_hgnc_cols}
\alias{.detect_hgnc_cols}
\title{Identify the HGNC columns and return the first
(Note - this is by looking for Gapdh which may not be the best metric...)}
\usage{
.detect_hgnc_cols(gpl.df)
}
\arguments{
\item{gpl.df}{the gpl data frame}
}
\value{
the index of the column containing these ids
}
\description{
Identify the HGNC columns and return the first
(Note - this is by looking for Gapdh which may not be the best metric...)
}
