% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapDetectGeneID.R
\name{.detect_int_cols}
\alias{.detect_int_cols}
\title{Identify integer only only columns, hopefully these are Entrez.
(that will be tested later)}
\usage{
.detect_int_cols(gpl.df, MIN.OVERLAP = 8000)
}
\arguments{
\item{gpl.df}{the gpl data frame}

\item{MIN.OVERLAP}{the minimum number of mapped genes allowed, default=8000}
}
\value{
a list of the indices of the columns containing these ids
}
\description{
Identify integer only only columns, hopefully these are Entrez.
(that will be tested later)
}
