% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapDetectGeneID.R
\name{.detect_unigene_cols}
\alias{.detect_unigene_cols}
\title{Identify the unigene columns and return the one with the most overalp}
\usage{
.detect_unigene_cols(gpl.df, MIN.OVERLAP = 8000)
}
\arguments{
\item{gpl.df}{the gpl data frame}

\item{MIN.OVERLAP}{the minimum number of mapped genes allowed, default=8000}
}
\value{
the index of the column containing these ids
}
\description{
Identify the unigene columns and return the one with the most overalp
}
