% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{.expDataToRanks}
\alias{.expDataToRanks}
\title{Convert an expression dataset to ranks.}
\usage{
.expDataToRanks(gene_mat)
}
\arguments{
\item{gene_mat}{an expression matrix with genes as rows and columns as samples}
}
\value{
rank_dat ranked dataset
}
\description{
Briefly, takes an expression matrix, and then ranks each column, 1 to n (number of genes)
in order of increasing expression values. Missing data is ranked last.
}
