% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapSetupRef.R
\name{.extract_biomart_ref}
\alias{.extract_biomart_ref}
\title{Extract a gene mapping table from biomaRt, this contains ensembl, refseq, entrez ids
and chromsome names. HGNC symbols are included for human and mouse.}
\usage{
.extract_biomart_ref(organism)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}
}
\value{
mapping table for use in converting between IDs
}
\description{
Extract a gene mapping table from biomaRt, this contains ensembl, refseq, entrez ids
and chromsome names. HGNC symbols are included for human and mouse.
}
