% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRefTables.R
\name{.extract_entrez}
\alias{.extract_entrez}
\title{Gets a list of all entrez ids for an organism.
This pulls the list from the gene map, which is either in the ref_dir or is generated.}
\usage{
.extract_entrez(organism, ref_dir)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}

\item{ref_dir}{the reference directory to pull from, this is either provided or a tempdir}
}
\value{
a list of entrez ids
}
\description{
Gets a list of all entrez ids for an organism.
This pulls the list from the gene map, which is either in the ref_dir or is generated.
}
