% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_lab_code.R
\name{.filterGenesByVar}
\alias{.filterGenesByVar}
\title{Helper function to filter genes based on their standard deviations.}
\usage{
.filterGenesByVar(expr_mat, genes, cut_frac = "1st Qu.")
}
\arguments{
\item{expr_mat}{expr_mat - matrix of expression genes}

\item{genes}{list of sex labeling genes}

\item{cut_frac}{the fraction}
}
\value{
a filtered set of sex labeling genes
}
\description{
Removes sex labeling genes with variance below the provided quartile cut. This can
be helpful for removing genes on a platform that have low variance,
and such are not informative for sex labeling.
}
