% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapDetectGeneID.R
\name{.find_entrez_w_in_block}
\alias{.find_entrez_w_in_block}
\title{Identify columns that might have entrez in them. This would be
a block column where the entrez column is separated by "///" or ","}
\usage{
.find_entrez_w_in_block(gpl.df, MIN.OVERLAP = 8000)
}
\arguments{
\item{gpl.df}{the gpl data frame}

\item{MIN.OVERLAP}{the minimum number of mapped genes allowed, default=8000}
}
\value{
a list of the indices of the columns containing these ids
}
\description{
Identify columns that might have entrez in them. This would be
a block column where the entrez column is separated by "///" or ","
}
