% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGeneMat.R
\name{.formatGeneProbe}
\alias{.formatGeneProbe}
\title{Take a list with probe to gene mapping and formate it as gene to probe mapping
 In the process, separate out to allow for multi-mapping}
\usage{
.formatGeneProbe(keys)
}
\arguments{
\item{keys}{a list probe to gene}
}
\value{
list with gene to probe mapping, genes are names, probes
}
\description{
Take a list with probe to gene mapping and formate it as gene to probe mapping
 In the process, separate out to allow for multi-mapping
}
