% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSexLab.R
\name{.geomMeanAcrossGenes}
\alias{.geomMeanAcrossGenes}
\title{Helper function to calculate geometric mean of a set of genes across all samples in a matrix.}
\usage{
.geomMeanAcrossGenes(expr_mat, genes)
}
\arguments{
\item{expr_mat}{expression dataset, rows are genes, columns are samples}

\item{genes}{list of genes to calculate the geometric mean of}
}
\value{
a vector of geometric means
}
\description{
Computes the geometric mean of the expression of a particular set of genes for every sample
(column) in a matrix. NAs are ignored during this calculation. The result is a vector of
geometric means, each corresponding to one sample (column) in the original matrix.
}
