% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRefTables.R
\name{.get_genbank}
\alias{.get_genbank}
\title{Gets a mapping of genbank to entrez.
Note - these tables are very large.}
\usage{
.get_genbank(organism)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}
}
\value{
a table with genbank and entrez ids
}
\description{
Gets a mapping of genbank to entrez.
Note - these tables are very large.
}
