% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMapSetupRef.R
\name{.get_ref_data}
\alias{.get_ref_data}
\title{Generate the desired mapping data for that organism}
\usage{
.get_ref_data(organism, type, ref_dir)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}

\item{type}{type of data desired}

\item{ref_dir}{the reference directory to pull from, this is either provided or a tempdir}
}
\value{
the desired mapping data
}
\description{
Generate the desired mapping data for that organism
}
