% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRefTables.R
\name{.get_unigene}
\alias{.get_unigene}
\title{Gets a mapping of unigene ids to entrez ids.
This pulls the list from the entrezids, which is either in the ref_dir or is generated.}
\usage{
.get_unigene(organism, ref_dir)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}

\item{ref_dir}{the reference directory to pull from, this is either provided or a tempdir}
}
\value{
table with unigene and entrez ids
}
\description{
Gets a mapping of unigene ids to entrez ids.
This pulls the list from the entrezids, which is either in the ref_dir or is generated.
}
