% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRefTables.R
\name{.get_xy_genes}
\alias{.get_xy_genes}
\title{Gets a table of all the X,Y genes for that organism with the entrez ids}
\usage{
.get_xy_genes(organism, ref_dir)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}

\item{ref_dir}{the reference directory to pull from, this is either provided or a tempdir}
}
\value{
a table with two columns "gene", the entrez gene id, and "chr",
       the chromosome name, which will be one of X and Y
}
\description{
Gets a table of all the X,Y genes for that organism with the entrez ids
}
