% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGeneMat.R
\name{.getGenes}
\alias{.getGenes}
\title{Get the list of genes for a GPL}
\usage{
.getGenes(gpl.name, ref_dir = NULL)
}
\arguments{
\item{ref_dir}{option to set reference directory  (defaults to tempdir)}
}
\value{
probe/gene data frame
}
\description{
Get the list of genes for a GPL
}
