% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGeneMat.R
\name{.getGeneToProbe}
\alias{.getGeneToProbe}
\title{Grab the gene to probe mapping for a given object}
\usage{
.getGeneToProbe(platform, gse_keys = NULL, platform_dir = NULL,
  ref_dir = NULL)
}
\arguments{
\item{platform}{the platform to download}

\item{gse_keys}{the keys from the gse_obj}

\item{platform_dir}{option to set platform dir  (defaults to tempdir)}

\item{ref_dir}{option to set reference directory  (defaults to tempdir)}
}
\value{
gene.to.probe a mapping from genes to probes
}
\description{
Grab the gene to probe mapping for a given object
}
