% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRefTables.R
\name{.load_ref}
\alias{.load_ref}
\title{Load the desired reference data mapping table.
The reference tables are used later for mapping gpl data to entrez ids.
This first checks if the mapping exists in the ref_dir and then if not,
calls a helper function to extract it.}
\usage{
.load_ref(organism, type, ref_dir = NULL, save_dat = TRUE)
}
\arguments{
\item{organism}{the organism to grab data for: rat, mouse, or human}

\item{type}{type of data desired - one of "ensembl", "refseq", "hgnc", "unigene", "genbank", or "entrezids"}

\item{ref_dir}{option to set reference directory  (defaults to tempdir)}

\item{save_dat}{whether to save the data in the reference directory, defaults to TRUE}
}
\value{
the desired mapping data
}
\description{
Load the desired reference data mapping table.
The reference tables are used later for mapping gpl data to entrez ids.
This first checks if the mapping exists in the ref_dir and then if not,
calls a helper function to extract it.
}
