% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplMap.R
\name{.map_from_hgnc}
\alias{.map_from_hgnc}
\title{Helper function to extract and map HGNC to entrez
Note - this does not work for rat.}
\usage{
.map_from_hgnc(gpl.df, my.col, organism, ref_dir = NULL)
}
\arguments{
\item{gpl.df}{the feature data from the GPL data frame}

\item{my.col}{the column containing the HGNC data}

\item{organism}{the organism: rat, mouse, or human}

\item{ref_dir}{option to set reference directory  (defaults to tempdir)}
}
\value{
a parsed probe/gene data frame
}
\description{
Helper function to extract and map HGNC to entrez
Note - this does not work for rat.
}
