% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectGeneID.R
\name{.parse_multi_col}
\alias{.parse_multi_col}
\title{Helper function to parse a pattern out of a multi-column.
Used for columns with many fields contained within and a consistent
defined pattern. This is an alternate to `.find_col_loc()` for cases
where all strings match a pattern (e.g. "ENSG") and there may be
any type of delimiter within the column.}
\usage{
.parse_multi_col(df, my.col, pattern)
}
\arguments{
\item{df}{the data.frame with feature data}

\item{my.col}{the column to examine}

\item{pattern}{the pattern to extract}
}
\value{
an expanded data frame with probe/gene data
}
\description{
Helper function to parse a pattern out of a multi-column.
Used for columns with many fields contained within and a consistent
defined pattern. This is an alternate to `.find_col_loc()` for cases
where all strings match a pattern (e.g. "ENSG") and there may be
any type of delimiter within the column.
}
