% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbaScore.R
\name{.predDf}
\alias{.predDf}
\title{Helper function to generate a predicted probability df for
a given data frame and fit that contains the scores and predictions}
\usage{
.predDf(fit, df)
}
\arguments{
\item{fit}{a sex labeling fit object}

\item{df}{the expression matrix, rows are genes and samples are columns}
}
\description{
Helper function to generate a predicted probability df for
a given data frame and fit that contains the scores and predictions
}
