% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{.reorderGeneMat}
\alias{.reorderGeneMat}
\title{Reorder the expression matrix so it contains the pre-specified list
of genes and fill in missing data with NAs.}
\usage{
.reorderGeneMat(gene_mat, gene_list = list_genes)
}
\arguments{
\item{gene_mat}{the expression matrix to reorder}

\item{gene_list}{the list of genes to use}
}
\value{
a reordered expression mat with NAs for missing genes
}
\description{
Reorder the expression matrix so it contains the pre-specified list
of genes and fill in missing data with NAs.
}
