% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_lab_code.R
\name{expDataToRanks}
\alias{expDataToRanks}
\title{Convert an expression dataset in probe form to ranks.}
\usage{
expDataToRanks(probe_mat, probe_map, list_genes = NULL)
}
\arguments{
\item{probe_mat}{an expression matrix with probes as rows and columns as samples}

\item{probe_map}{list mapping from probes to genes, names are probes, values are genes}

\item{list_genes}{list of all genes to extract, if not provided will use default list}
}
\value{
rank_dat ranked dataset with rows as genes
}
\description{
Briefly, takes an expression matrix (rows are probes) and mapping to genes,
and converts the matrix to genes, and then ranks each column, 1 to n (number of genes)
in order of decreasing expression data. Missing data is ranked last.
}
