% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprsex.R
\docType{package}
\name{exprsex}
\alias{exprsex}
\alias{exprsex-package}
\title{exprsex: A package for sex labeling microarray expression data.}
\description{
This package uses performs sex labeling of microarray expression datasets by comparing
the ranks of genes with consistent male- and female-specific expression in order
label samples as male or females. This method allows for robust sex-labeling
across a wide range of platforms and study sex breakdowns.
}
\section{exprsex functions}{

The exprsex functions are:
  predSexLab
}

