% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{getConsensusGenes}
\alias{getConsensusGenes}
\title{Create a consensus gene list from a set of expression matrices.
The goal of this function is to get a unified list of genes across studies.}
\usage{
getConsensusGenes(list.dats, min.fraction = 0.6, min.genes = 8000,
  na.max = 0.3)
}
\arguments{
\item{min.fraction}{the fraction number of studies that we allow, defaults to 0.6}

\item{min.genes}{the minimum number of genes that the studies need to contain to be considered
(if you use the package GPL mapping, it will already filter to a minimum threshold,
but use this to adjust further)}

\item{na.max}{maximum NA fraction to keep a row, defaults to 0.3}

\item{list.exprs}{the list of dataset objects from prepFromGSE or prepFromExpr}
}
\value{
list of genes present in min.fraction of the studies
}
\description{
This consensus gene list can then be used for reordering and ranking a dataset.
}
