% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGPL.R
\name{map_mult_gpl}
\alias{map_mult_gpl}
\title{Wrapper to parse and map data from multiple gpls.}
\usage{
map_mult_gpl(gpl.list, ref_dir = NULL, parallelize = FALSE)
}
\arguments{
\item{gpl.list}{list of platforms to map}

\item{ref_dir}{optional reference directory, defaults to a temporary directory}

\item{parallelize}{optionally parallelize the mapping}
}
\description{
Wrapper to parse and map data from multiple gpls.
}
