% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbaScore.R
\name{predProba}
\alias{predProba}
\title{Return the predicted probabilities and sex labels for an expression matrix.}
\usage{
predProba(fit = NULL, df)
}
\arguments{
\item{fit}{a sex labeling fit object with a logisitic fit}

\item{df}{the expression matrix, rows are genes and samples are columns}
}
\value{
a data frame with a column for samples, sex labels, and proba scores
}
\description{
Return the predicted probabilities and sex labels for an expression matrix.
}
