% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSexLab.R
\name{predSexLab}
\alias{predSexLab}
\title{Use a particular sex labeling fit to label an input dataset}
\usage{
predSexLab(fit, expr_mat, numeric_lab = TRUE)
}
\arguments{
\item{fit}{a sex labeling fit object}

\item{expr_mat}{data frame to label - rows are genes (HGNC_symbol), columns are samples, values are ranks}

\item{numeric_lab}{if FALSE, the output is "female/male", otherwise it is numeric (0/1)}
}
\value{
sex_lab - a list of sex labels (0/1 or "female"/"male")
}
\description{
Use a particular sex labeling fit to label an input dataset
}
