% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{prepFromExpr}
\alias{prepFromExpr}
\title{Prepare data from an existing expression matrix}
\usage{
prepFromExpr(expr.mat, mapping.mat = NULL, to.genes = FALSE,
  platform.id = NULL, gpl.dir = NULL, pheno = NULL)
}
\arguments{
\item{expr.mat}{expression matrix}

\item{mapping.mat}{the matrix of, defaults to NULL}

\item{to.genes}{whether the data needs to be converted to genes, defaults to FALSE
if this is TRUE, either the mapping matrix or platform.id need to be provided}

\item{platform.id}{the platform ID, defaults to NULL}

\item{gpl.dir}{directory where GPL ref data is location, empty defaults to tempdir()}

\item{pheno}{pheno data frame to add if desired, defaults to NULL}
}
\value{
list of gene expression matrices, with rows as genes and columns as samples
        values are gene expression levels
}
\description{
Prepare data from an existing expression matrix
}
