% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{prepInput}
\alias{prepInput}
\title{Prepare the input data for sex labeling.}
\usage{
prepInput(probe_mat, probe_map, gene_list = list_genes)
}
\arguments{
\item{probe_mat}{an expression matrix with probes as rows and columns as samples}

\item{probe_map}{list mapping from probes to genes, names are probes, values are genes}

\item{gene_list}{list of all genes to extract, if not provided will use default list}
}
\value{
rank_dat ranked dataset with rows as genes
}
\description{
Briefly, takes an expression matrix with rows
}
