% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_lab_code.R
\name{runSexLab}
\alias{runSexLab}
\title{Run sex labeling using default settings.}
\usage{
runSexLab(expr_mat, platform = NULL, is_rank = FALSE,
  numeric_lab = FALSE)
}
\arguments{
\item{expr_mat}{expression data matrix - rows are genes, columns are samples}

\item{platform}{the microarray platform (e.g. GPL570), if specified will use a model specific to that platform}

\item{is_rank}{TRUE if the data is already in rank form, otherwise will perform the conversion}
}
\value{
sex_lab - a list of sex labels (0/1 or "female"/"male")
}
\description{
Run sex labeling using default settings.
}
