% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex_lab_genes}
\alias{sex_lab_genes}
\title{Default genes used for sex labeling.}
\format{An object of class \code{list} of length 2.}
\usage{
sex_lab_genes
}
\description{
A list containing two attributes: "f" and "m" which are the female and male-specific genes respectively.
These genes were identified using meta-analysis, and are used to sex label data unless other genes
are provided.
}
\keyword{datasets}
