% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbaScore.R
\name{trainProba}
\alias{trainProba}
\title{Fit a logistic curve to the sex labeling fit so that we can get
accurate proba scores for the output.
Note that it is possible (and encouraged!) to use a representative
but different dataset to fit the curve as used to generate the SL object.}
\usage{
trainProba(fit = NULL, df, labels)
}
\arguments{
\item{fit}{a sex labeling fit object}

\item{df}{the expression matrix, rows are genes and samples are columns}

\item{labels}{sex labels in numeric (0/1) form for the expression mat}
}
\description{
Fit a logistic curve to the sex labeling fit so that we can get
accurate proba scores for the output.
Note that it is possible (and encouraged!) to use a representative
but different dataset to fit the curve as used to generate the SL object.
}
