% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSexLab.R
\name{trainSexLab}
\alias{trainSexLab}
\title{Train a sex labeling classifier using input training data.}
\usage{
trainSexLab(train_dat, train_lab, female_genes = NULL,
  male_genes = NULL, cut_frac = "1st Qu.")
}
\arguments{
\item{train_dat}{training data frame - rows are genes (entrez id), columns are samples, values are ranks}

\item{train_lab}{named list of training labels - 0 for female, 1 for male}

\item{female_genes}{a list of female-specific genes (entrez id), if not provided - default genes will be used}

\item{male_genes}{a list of male-specific genes (entrez id), if not provided, default genes will be used}

\item{cut_frac}{}
}
\value{
fit
}
\description{
Train a sex labeling classifier using input training data.
}
