% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeXYplot.R
\name{visualizeXYplot}
\alias{visualizeXYplot}
\title{This function graphs the XIST vs. RPS4Y1 probes for a given study.}
\usage{
visualizeXYplot(expr_matrix, keys, sex_lab, x_axis = "XIST",
  y_axis = "RPS4Y1")
}
\arguments{
\item{expr_matrix}{expression data matrix - rows are genes, columns are samples}

\item{keys}{mapping between genes and probes}

\item{sex_lab}{sex labels (0 - female, 1 - male) for each of the samples}
}
\value{
a graph of RPS4Y1 vs. XIST probes corresponding to the sex labels of studies
}
\description{
This function graphs the XIST vs. RPS4Y1 probes for a given study.
}
