% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{.checkExprMatFormat}
\alias{.checkExprMatFormat}
\title{Checks that an input expression matrix has the correct format}
\usage{
.checkExprMatFormat(expr_mat)
}
\arguments{
\item{expr_mat}{the expression matrix}
}
\description{
Makes sure the object is a data.frame or matrix, and the values in it are numeric.
}
