% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGPL.R
\name{.map_from_genbank}
\alias{.map_from_genbank}
\title{Helper function to extract and map genbank to entrez}
\usage{
.map_from_genbank(gpl.df, my.col, organism, ref_dir = NULL,
  col.given = FALSE)
}
\arguments{
\item{gpl.df}{the feature data from the GPL data frame}

\item{my.col}{the column containing the genbank data}

\item{organism}{the organism: rat, mouse, or human}

\item{ref_dir}{option to set reference directory  (defaults to tempdir)}

\item{col.given}{whether we know the column by name --> don't find loc (defaults to FALSE)}
}
\value{
a parsed probe/gene data frame
}
\description{
Helper function to extract and map genbank to entrez
}
