% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{getPrepGSE}
\alias{getPrepGSE}
\title{Get and prepare a gse file for sex labeling by converting it
to genes. Optionally this will also reorder by a list of genes and
convert it to ranks.}
\usage{
getPrepGSE(gse, to.ranks = FALSE, gse.dir = NULL, gpl.dir = NULL,
  out.dir = NULL, gene_list = NULL)
}
\arguments{
\item{gse}{the ID of the GSE to download}

\item{to.ranks}{whether to convert to ranks, default is false}

\item{gse.dir}{directory that contains gse files, if empty, will download the GSE}

\item{gpl.dir}{directory where GPL ref data is location, empty defaults to tempdir()}

\item{out.dir}{directory to write output, if empty will just be returned and not be written out}

\item{gene_list}{the list of genes to use, this is helpful if you want to compare across multiple studies}
}
\value{
gene expression matrix, with rows as genes and columns as samples
        values are gene expression levels unless converted to ranks
}
\description{
Get and prepare a gse file for sex labeling by converting it
to genes. Optionally this will also reorder by a list of genes and
convert it to ranks.
}
