% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGPL.R
\name{parse_entrez_from_gpl}
\alias{parse_entrez_from_gpl}
\title{Parse and map entrez data from a GPL annotation.}
\usage{
parse_entrez_from_gpl(gpl.name, ref_dir = NULL, MIN.OVERLAP = 8000)
}
\arguments{
\item{gpl.name}{the name gpl to load}

\item{ref_dir}{optional reference directory, defaults to a temporary directory}

\item{MIN.OVERLAP}{the minimum number of mapped genes allowed, default is 8000}
}
\value{
a data frame with "probe" and "gene" columns, where gene is the entrezid
}
\description{
Parse and map entrez data from a GPL annotation.
}
