% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_lab_code.R
\name{predSexLab}
\alias{predSexLab}
\title{Use a particular sex labeling fit to label an input dataset}
\usage{
predSexLab(fit, expr_mat, numeric_lab = FALSE)
}
\arguments{
\item{fit}{a sex labeling fit object}

\item{expr_mat}{data frame to label - rows are genes (HGNC_symbol), columns are samples, values are ranks}

\item{numeric_lab}{if TRUE output is numeric (0/1), otherwise it is ("female"/"male")}
}
\value{
sex_lab - a list of sex labels (0/1 or "female"/"male")
}
\description{
Use a particular sex labeling fit to label an input dataset
}
