% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{prepInput}
\alias{prepInput}
\title{Prepare the input data for sex labeling by converting it to genes
and getting the ranks.}
\usage{
prepInput(gse.obj, gene_list = list_genes)
}
\arguments{
\item{gse.obj}{the object}

\item{gene_list}{the list of genes to use}
}
\value{
the gse.obj prepared for sex labeling
}
\description{
Prepare the input data for sex labeling by converting it to genes
and getting the ranks.
}
